AP-10331: Add compatibilityMode on Java side
Note that a new method was needed in DLPythonNetworkLoader that takes
the network instead of only the network source because the network is
needed to check if the compatibility mode should be enabled.

Also note, that the check if loading the network with training
configuration fails (and trying it again without) was moved to the keras
integration because it is keras specific.